+++ docs/release_notes.md
@@ -17,6 +17,9 @@
 
 ### ✨ New Functionality
 
+- [OpenAI] Support for OpenAI Realtime API with the new OpenAiRealtimeClient. The following cases are currently supported:
+  - `textToSpeech`
+  - `speechToSpeech`
 - [Orchestration] Added `QWEN_3_7_MAX`, `QWEN_3_7_PLUS` and `ALIBABA_TEXT_EMBEDDING_4` to model list in `OrchestrationAiModel`.
 - [Orchestration] Marked as deprecated `QWEN_3_MAX` model (retired, use newer models instead, e.g. `QWEN_3_7_MAX`).
 - [OpenAI] You can now add multiple custom headers to an `OpenAiClient` at once via `.withHeaders()`.